Update Français.txt
Hello ! 
I have a little question with the translation. 
There are a lot of times the word "Room" to translate, the problem is that in french this word can have a lot of sense. I know this game so I know that there are only 2 different translations for french but I don't know wich one I have to use.
Sometimes Room means Bedromm in french, and sometimes it juste means a place. 

The translation can't be totally correct if I don't know with what "room" is linked, is it a special place like the bedroom, or just a general place of the hotel?

This is just what I wanted to say to help more.